Cleanup and Config File
Cleaned up the code a little, and also created a config file called paths.txt in main directory to read where each file should be and where each corresponding file should go to. This is in order to release this code to Public GitHub and just enables more flexibility in the longrun when changing code.

Signed-off-by: n30w <99268950+n30w@users.noreply.github.com>