Type Hints + Links
Look, I'm new to type hints. I don't know how they work, so I kind of splattered them all over the place in the code. I think they're helpful since I'm usually a Java programmer and that thing is so strict when it comes to types. Anyway, I may improve these type hints later.

I also reformatted and added more helpful links to the README.md. I had trouble with the venv on M1 as well, so I kind of added a tutorial to that.

Signed-off-by: n30w <99268950+n30w@users.noreply.github.com>